+++ packages/editor-common/web/src/index.ts
@@ -44,6 +44,7 @@ export {
   getAnchorBlockData,
   mergeBlockData,
   isAtomicBlockFocused,
+  blockKeyToEntityKey,
   setEntityData,
   setBlockNewEntityData,
   replaceWithEmptyBlock,
@@ -60,6 +61,7 @@ export {
   getPostContentSummary,
   createSelection,
   getBlockType,
+  hasInlineStyle,
   indentSelectedBlocks,
   isTypeText,
   setForceSelection,
@@ -77,6 +79,7 @@ export {
   isCursorAtFirstLine,
   selectAllContent,
 } from './Utils/draftUtils';
+export { triggerMention, insertMention } from './Utils/mentionUtils';
 export { isiOS } from './Utils/isiOS';
 export { mergeToolbarSettings } from './Utils/mergeToolbarSettings';
 export {