+++ CHANGELOG.md
@@ -16,10 +16,19 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `common`
+  - [#2344](https://github.com/wix-incubator/rich-content/pull/2344) fix tooltip z-index for media settings modal
+
 </details>
 <hr/>
 
-## 8.29.2 (April 5, 2021)
+## 8.29.2 (April 6, 2021)
+### :rocket: New Feature
+- `ricos-editor`
+  - [#2044](https://github.com/wix-incubator/rich-content/pull/2044) **beta** Editor Commands
+
 ### :nail_care: Polish
 - `image`
   - [#2343](https://github.com/wix-incubator/rich-content/pull/2343) skip LCP optimization for png and mobile